content/media/webaudio/test/test_singleSourceDest.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test whether we can create an AudioContext interface</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="webaudio.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  SpecialPowers.setBoolPref("media.webaudio.enabled", true);

  var context = new AudioContext();
  var buffer = context.createBuffer(1, 2048, context.sampleRate);
  for (var i = 0; i < 2048; ++i) {
    buffer.getChannelData(0)[i] = Math.sin(440 * 2 * Math.PI * i / context.sampleRate);
  }

  var destination = context.destination;
  is(destination.context, context, "Destination node has proper context");
  is(destination.context, context, "Destination node has proper context");
  is(destination.numberOfInputs, 1, "Destination node has 1 inputs");
  is(destination.numberOfOutputs, 0, "Destination node has 0 outputs");

  testWith(context, buffer, destination, function(source) {
    source.start(0);
  }, function(source) {
    source.stop();
  }, function() {
    testWith(context, buffer, destination, function(source) {
      source.start(0, 1);
    }, function(source) {
      expectTypeError(function() {
        source.noteOff();
      });
      source.noteOff(0);
    }, function() {
      testWith(context, buffer, destination, function(source) {
        source.start(0, 1, 0.5);
      }, function(source) {
        source.stop(0);
      }, function() {
        testWith(context, buffer, destination, function(source) {
          source.noteOn(0);
        }, function(source) {
          source.noteOff(0);
        }, function() {
          testWith(context, buffer, destination, function(source) {
            source.noteGrainOn(0, 1, 0.5);
          }, function(source) {
            source.stop();
          }, function() {
            SpecialPowers.clearUserPref("media.webaudio.enabled");
            SimpleTest.finish();
          });
        });
      });
    });
  });
});

function testWith(context, buffer, destination, start, stop, callback)
{
  var source = createNode(context, buffer, destination);
  start(source);
  SimpleTest.executeSoon(function() {
    stop(source);
    callback();
    source.disconnect();
  });
}

function createNode(context, buffer, destination) {
  var source = context.createBufferSource();
  is(source.context, context, "Source node has proper context");
  is(source.numberOfInputs, 0, "Source node has 0 inputs");
  is(source.numberOfOutputs, 1, "Source node has 1 outputs");
  is(source.loop, false, "Source node is not looping");
  is(source.loopStart, 0, "Correct default value for loopStart");
  is(source.loopEnd, 0, "Correct default value for loopEnd");
  ok(!source.buffer, "Source node should not have a buffer when it's created");

  source.buffer = buffer;
  ok(source.buffer, "Source node should have a buffer now");

  source.connect(destination);

  is(source.numberOfInputs, 0, "Source node has 0 inputs");
  is(source.numberOfOutputs, 1, "Source node has 0 outputs");
  is(destination.numberOfInputs, 1, "Destination node has 0 inputs");
  is(destination.numberOfOutputs, 0, "Destination node has 0 outputs");

  return source;
}

</script>
</pre>
</body>
</html>
